feat: added logs and exceptions

+++ src/main/java/com/challenge/geosapiens/service_b/infrastructure/usecase/order/DeleteOrderUseCaseImpl.java
@@ -1,5 +1,7 @@
 package com.challenge.geosapiens.service_b.infrastructure.usecase.order;
 
+import com.challenge.geosapiens.service_b.application.exception.NotFoundException;
+import com.challenge.geosapiens.service_b.domain.entity.Order;
 import com.challenge.geosapiens.service_b.domain.repository.OrderRepository;
 import com.challenge.geosapiens.service_b.domain.usecase.order.DeleteOrderUseCase;
 import lombok.RequiredArgsConstructor;
@@ -19,10 +21,15 @@ public class DeleteOrderUseCaseImpl implements DeleteOrderUseCase {
     @Override
     @Transactional
     public void execute(UUID orderId) {
-        log.debug("Deleting order with ID: {}", orderId);
+        log.info("[DeleteOrderUseCaseImpl] Starting order deletion with ID: {}", orderId);
+
+        if(!orderRepository.existsById(orderId)) {
+            log.error("[DeleteOrderUseCaseImpl] Order not found with ID: {}", orderId);
+            throw new NotFoundException("Order not found with ID: " + orderId);
+        }
 
         orderRepository.deleteById(orderId);
 
-        log.info("Order deleted successfully with ID: {}", orderId);
+        log.info("[DeleteOrderUseCaseImpl] Order deleted successfully with ID: {}", orderId);
     }
 }
